src/orchestrator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
numpy==1.26.2
loguru==0.7.2
pyyaml==6.0.1
python-multipart==0.0.6
